created useFetchMetrics.js hook
