+++ ex-2/dashboard-data-service.js
@@ -1,49 +1,10 @@
 import { useState } from 'react';
+import useFetchMetrics from './use-fetch-metrics';
 
 const DashboardDataService = () => {
-	let isLoading = false;
-	let cache = {};
-	let lastUpdated = null;
-	
 
 	// BUG: This function has memory leaks and doesn't handle errors properly
-	const fetchMetrics = async ( dateRange ) => {
-		isLoading = true;
-		
-		// Create endpoint URL
-		let endpoint = '/api/metrics?';
-		endpoint += 'start=' + dateRange.start;
-		endpoint += '&end=' + dateRange.end;
-		
-		try {
-			const cacheKey = `${dateRange.start}${dateRange.end}`;
-			const cache = cache[cacheKey];
-			if ( cache ) {
-				return cache;
-			}
-
-			const response = await fetch( endpoint );
-			const data = await response.json();
-
-			const transformed = data.metrics.map( metric => ( {
-					name: metric.name,
-					value: metric.value,
-					change: metric.previous ? ( metric.value - metric.previous ) / metric.previous : 0,
-					trend: metric.trend
-			} ) );
-
-
-			cache[cacheKey] = transformed;
-			lastUpdated = new Date();
-			return transformed;
-		} catch ( error ) {
-			// Poor error handling
-			console.log( 'Error fetching metrics:', error );
-			return [];
-		} finally {
-			isLoading = false;
-		}
-	}
+	const fetchMetrics = useFetchMetrics( dateRange );
 
 	// BUG: This function doesn't aggregate data correctly
 	const getAggregatedData = ( metrics, groupBy ) => {
